requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.1
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
django-environ==0.8.1

django-smart-selects==1.5.9
django-ckeditor==6.4.0
mssql-django==1.1.2

django_google_maps==0.13.0
django_seed==0.3.1
reportlab==3.6.11